+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/findreplace/FindReplaceBar.java
@@ -136,6 +136,33 @@ public class FindReplaceBar extends Composite implements Display, RequiresResize
 
       shelf.addLeftWidget(panel);
 
+
+      if (RStudioGinjector.INSTANCE.getUserPrefs().findPanelLegacyTabSequence().getValue())
+      {
+         // RStudio 1.2 and earlier set custom tabindexes so hitting Tab key would go from the
+         // "find" text to the "replace" text and skip some of the other controls. This was
+         // a keyboard accessibility violation, but some users would like option to get that
+         // old behavior back. https://github.com/rstudio/rstudio/issues/7295
+
+         // fixup tab indexes of controls
+         txtFind_.setTabIndex(100);
+         txtReplace_.setTabIndex(101);
+         chkInSelection_.setTabIndex(102);
+         chkCaseSensitive_.setTabIndex(103);
+         chkWholeWord_.setTabIndex(104);
+         chkRegEx_.setTabIndex(105);
+         chkWrapSearch_.setTabIndex(106);
+
+         // remove SmallButton instances from tab order since (a) they aren't
+         // capable of showing a focused state; and (b) enter is already a
+         // keyboard shortcut for both find and replace
+         btnFindNext_.setTabIndex(-1);
+         btnFindPrev_.setTabIndex(-1);
+         btnSelectAll_.setTabIndex(-1);
+         btnReplace_.setTabIndex(-1);
+         btnReplaceAll_.setTabIndex(-1);
+      }
+
       shelf.setRightVerticalAlignment(HasVerticalAlignment.ALIGN_TOP);
       shelf.addRightWidget(btnClose_ = new Button());
       btnClose_.setStyleName(RES.styles().closeButton());